restructure components in info folder
