+++ angular/src/app/app.module.ts
@@ -10,13 +10,14 @@ import { TagCloudModule } from 'angular-tag-cloud-module';
 import { CorrelationService } from './data/correlation.service';
 import { KeywordMappingService } from './keyword-mapping/keyword-mapping.service';
 import { AnalogyService } from './data/analogy.service';
-import { InfoService } from './info/info.service';
+import { InfoService } from './info/data/info.service';
 
 import { AppComponent } from './app.component';
 import { HomeComponent} from './home/home.component';
-import { CollectionInfoComponent } from './info/collection-info.component';
 import { InfoNavigationComponent } from './info/info-navigation.component';
-import { ModelInfoComponent } from './info/model-info.component';
+import { CollectionInfoComponent } from './info/collection-info/collection-info.component';
+import { ModelInfoComponent } from './info/model-info/model-info.component';
+import { ReliabilityComponent } from './info/reliability/reliability.component';
 import { ExploreComponent } from './explore/explore.component';
 import { ManualComponent } from './manual/manual.component';
 import { AboutComponent } from './about/about.component';
@@ -26,19 +27,19 @@ import { WordcloudComponent } from './wordcloud/wordcloud.component';
 import { AnalogiesComponent } from './analogies/analogies.component';
 import { ModelSelectionComponent } from './model-selection/model-selection.component';
 import { DropdownDirective } from './shared/dropdown.directive';
-import { ReliabilityComponent } from './info/reliability.component';
 
 const appRoutes: Routes = [
-  { path: '', component: HomeComponent },
+  { path: 'home', component: HomeComponent },
   { path: 'manual', component: ManualComponent },
   { path: 'about', component: AboutComponent },
   { path: 'explore', component: ExploreComponent,
 		children: [
-			{path: '', component: CollectionInfoComponent, outlet:'info'},
+			{path: '', redirectTo: '/explore/(info:collection)', pathMatch: 'full'},
 			{path: 'model', component: ModelInfoComponent, outlet:'info'},
 			{path: 'collection', component: CollectionInfoComponent, outlet:'info'},
 			{path: 'reliability', component: ReliabilityComponent, outlet:'info'},
-		]}
+		]},
+  { path: '', redirectTo: '/home', pathMatch: 'full'},
 ];
