+++ angular/src/app/info/data/info.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 
 import { ModelInfo } from './model-info.interface';
 import { CollectionInfo } from './collection-info.interface';
