+++ angular/src/app/info/data/reliability.interface.ts
