+++ angular/src/app/info/reliability/reliability.component.css
